+++ .drone.yml
@@ -42,7 +42,7 @@ steps:
         go test
         -failfast
         -timeout=20m
-        -tags "wasmsqlite3 netgo osusergo static_build kvformat timetzdata"
+        -tags "netgo osusergo static_build kvformat timetzdata"
         ./...
       - ./test/envparsing.sh
       - ./test/swagger.sh
